Allow import comments to be inherited.
When processing a subpackage of a package which has an import comment, if that
subpackage does not have its own import comment then try to derive one from the
nearest enclosing parent package.

This ensures repositories with incomplete comment coverage will get sensible
results. The build tool has an implicit path dependency and will not be fooled
by this since it has in hand the desired import path: Post facto analysis tools
like this do not have that context.