deps: Check for module names while assigning import paths.
Although we don't fully support modules (in particular, we do not do anything
to harmonize the repository state with the version targeted by the import),
this ensures we at least get the right import paths most of the time.

The module file parsing is basic and hacky, but the official parser is still
locked behind an internal and is too much code to bother cloning for this
simple a purpose.